fix(client/http): error when used in context without Segment
When the client was used in a code path without a Segment, the creation
of the Subsegment failed but the error was silently caught (as intended).

In the later interceptor functions, the client crashed when accessing the
missing property config[TRACING_CONFIG_KEY]. By explicitly checking if
it is set, we avoid this bug.