feat(core): allow setting custom XRay plugins
We hard-coded the EC2 and ECS plugins because we needed those at narando.

By making it configurable the plugin can be more easily adopted in other
environments without overcomplicating things.

We will keep the EC2 and ECS as the default plugins to not break
compatibility, but this will change with the next major version.