fix(env/http): segment has invalid URL when sampling decision is made
By applying the middleware to path pattern `"*"`, express will replace the
`req.url` property with `"/"`, we tried to fix this by setting the correct
url after the segment was created and the callback from the XRay middleware
is called.

We can avoid this workaround by registering the middleware to `"/"`, this
way express will not replace any part of `req.url`.

A similar error from nestjs/nest that describes a similar issue and why
express does it that way: https://github.com/nestjs/nest/issues/4315